projekt_python_back/requirements.txt
Flask
flask-mysqldb
flask-cors
mysql-connector-python